Move to Spark 4.0
