+++ api/src/main/java/org/apache/iceberg/types/Comparators.java
@@ -48,6 +48,7 @@ public class Comparators {
           .put(Types.StringType.get(), Comparators.charSequences())
           .put(Types.UUIDType.get(), Comparator.naturalOrder())
           .put(Types.BinaryType.get(), Comparators.unsignedBytes())
+          .put(Types.UnknownType.get(), Comparator.naturalOrder())
           .buildOrThrow();
 
   public static Comparator<StructLike> forType(Types.StructType struct) {
