+++ arrow/src/main/java/org/apache/iceberg/arrow/ArrowSchemaUtil.java
@@ -134,6 +134,9 @@ public class ArrowSchemaUtil {
             new Field("", new FieldType(field.isOptional(), arrowType, null), entryFields);
         children.add(entry);
         break;
+      case UNKNOWN:
+        arrowType = new ArrowType.Null();
+        break;
       default:
         throw new UnsupportedOperationException("Unsupported field type: " + field);
     }
