+++ spark/v3.5/spark/src/main/java/org/apache/iceberg/spark/TypeToSparkType.java
@@ -38,7 +38,6 @@ import org.apache.spark.sql.types.LongType$;
 import org.apache.spark.sql.types.MapType$;
 import org.apache.spark.sql.types.Metadata;
 import org.apache.spark.sql.types.MetadataBuilder;
-import org.apache.spark.sql.types.NullType$;
 import org.apache.spark.sql.types.StringType$;
 import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType$;
@@ -125,8 +124,6 @@ class TypeToSparkType extends TypeUtil.SchemaVisitor<DataType> {
       case DECIMAL:
         Types.DecimalType decimal = (Types.DecimalType) primitive;
         return DecimalType$.MODULE$.apply(decimal.precision(), decimal.scale());
-      case UNKNOWN:
-        return NullType$.MODULE$;
       default:
         throw new UnsupportedOperationException(
             "Cannot convert unknown type to Spark: " + primitive);
