+++ spark/v3.5/spark/src/test/java/org/apache/iceberg/spark/data/AvroDataTestBase.java
@@ -32,6 +32,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.iceberg.MetadataColumns;
 import org.apache.iceberg.Schema;
@@ -120,10 +121,15 @@ public abstract class AvroDataTestBase {
 
   @Test
   public void testStructWithRequiredFields() throws IOException {
+    List<Types.NestedField> supportedPrimitivesExceptUnknown =
+        SUPPORTED_PRIMITIVES.fields().stream()
+            .filter(f -> !f.type().typeId().equals(Type.TypeID.UNKNOWN))
+            .collect(Collectors.toList());
+
     writeAndValidate(
         TypeUtil.assignIncreasingFreshIds(
             new Schema(
-                Lists.transform(SUPPORTED_PRIMITIVES.fields(), Types.NestedField::asRequired))));
+                Lists.transform(supportedPrimitivesExceptUnknown, Types.NestedField::asRequired))));
   }
 
   @Test
@@ -602,49 +608,5 @@ public abstract class AvroDataTestBase {
                     33L)),
             record.copy(Map.of("id", 4L, "data", "d", "_row_id", 1_001L)),
             record.copy(Map.of("id", 5L, "data", "e"))));
-  }
-
-  @Test
-  public void testUnknownNestedLevel() throws IOException {
-    assumeThat(supportsNestedTypes()).isTrue();
-
-    Schema schema =
-        new Schema(
-            required(1, "id", LongType.get()),
-            optional(
-                2,
-                "nested",
-                Types.StructType.of(
-                    required(20, "int", Types.IntegerType.get()),
-                    optional(21, "unk", Types.UnknownType.get()))));
-
-    writeAndValidate(schema);
-  }
-
-  @Test
-  public void testUnknownListType() throws IOException {
-    assumeThat(supportsNestedTypes()).isTrue();
-
-    Schema schema =
-        new Schema(
-            required(0, "id", LongType.get()),
-            optional(1, "data", ListType.ofOptional(2, Types.UnknownType.get())));
-
-    writeAndValidate(schema);
-  }
-
-  @Test
-  public void testUnknownMapType() throws IOException {
-    assumeThat(supportsNestedTypes()).isTrue();
-
-    Schema schema =
-        new Schema(
-            required(0, "id", LongType.get()),
-            optional(
-                1,
-                "data",
-                MapType.ofOptional(2, 3, Types.StringType.get(), Types.UnknownType.get())));
-
-    writeAndValidate(schema);
   }
 }
