+++ spark/v3.5/spark/src/test/java/org/apache/iceberg/spark/data/TestSparkOrcReader.java
@@ -21,8 +21,6 @@ package org.apache.iceberg.spark.data;
 import static org.apache.iceberg.spark.data.TestHelpers.assertEquals;
 import static org.apache.iceberg.types.Types.NestedField.required;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.Assumptions.assumeThat;
 
 import java.io.File;
 import java.io.IOException;
@@ -108,23 +106,5 @@ public class TestSparkOrcReader extends AvroDataTestBase {
 
   private Iterator<InternalRow> batchesToRows(Iterator<ColumnarBatch> batches) {
     return Iterators.concat(Iterators.transform(batches, ColumnarBatch::rowIterator));
-  }
-
-  @Test
-  @Override
-  public void testUnknownListType() {
-    assumeThat(supportsNestedTypes()).isTrue();
-
-    assertThatThrownBy(super::testUnknownListType)
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessageStartingWith("Cannot create ListType with unknown element type");
-  }
-
-  @Test
-  @Override
-  public void testUnknownMapType() {
-    assertThatThrownBy(super::testUnknownMapType)
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessageStartingWith("Cannot create MapType with unknown value type");
   }
 }
