+++ spark/v3.5/spark/src/test/java/org/apache/iceberg/spark/data/TestSparkRecordOrcReaderWriter.java
@@ -20,7 +20,6 @@ package org.apache.iceberg.spark.data;
 
 import static org.apache.iceberg.types.Types.NestedField.required;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.io.File;
 import java.io.IOException;
@@ -150,21 +149,5 @@ public class TestSparkRecordOrcReaderWriter extends AvroDataTestBase {
     }
     assertThat(expectedIter).as("Expected iterator should not have any extra rows.").isExhausted();
     assertThat(actualIter).as("Actual iterator should not have any extra rows.").isExhausted();
-  }
-
-  @Test
-  @Override
-  public void testUnknownListType() {
-    assertThatThrownBy(super::testUnknownListType)
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessageStartingWith("Cannot create ListType with unknown element type");
-  }
-
-  @Test
-  @Override
-  public void testUnknownMapType() {
-    assertThatThrownBy(super::testUnknownMapType)
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessageStartingWith("Cannot create MapType with unknown value type");
   }
 }
