+++ spark/v4.0/spark/src/main/java/org/apache/iceberg/spark/PruneNullType.java
