+++ spark/v4.0/spark/src/test/java/org/apache/iceberg/spark/data/TestSparkParquetReader.java
@@ -244,4 +244,20 @@ public class TestSparkParquetReader extends AvroDataTestBase {
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessage("Missing required field: missing_str");
   }
+
+  @Test
+  @Override
+  public void testUnknownListType() {
+    assertThatThrownBy(super::testUnknownListType)
+        .isInstanceOf(IllegalArgumentException.class)
+        .hasMessageStartingWith("Cannot convert element Parquet: unknown");
+  }
+
+  @Test
+  @Override
+  public void testUnknownMapType() {
+    assertThatThrownBy(super::testUnknownMapType)
+        .isInstanceOf(IllegalArgumentException.class)
+        .hasMessageStartingWith("Cannot convert value Parquet: unknown");
+  }
 }
